Move config classes to separate files
