+++ Context/NpcContext.cs
@@ -3,6 +3,7 @@ using Microsoft.Xna.Framework.Graphics;
 using System.Collections.Generic;
 using Terraria;
 using Terraria.Localization;
+using Twaila.Config;
 using Twaila.Graphics;
 using Twaila.Systems;
 using Twaila.UI;
@@ -62,7 +63,7 @@ namespace Twaila.Context
 
         public override void Update()
         {
-            TwailaConfig.Content content = TwailaConfig.Instance.DisplayContent;
+            Content content = TwailaConfig.Instance.DisplayContent;
 
             IntersectsNPC(Pos.BestPos(), out NPC foundNPC);
