+++ Context/PalmTreeContext.cs
@@ -5,6 +5,7 @@ using Terraria.ModLoader;
 using Twaila.Util;
 using Twaila.Graphics;
 using Twaila.Systems;
+using Twaila.Config;
 
 namespace Twaila.Context
 {
