+++ Util/TileUtil.cs
@@ -5,6 +5,7 @@ using Terraria;
 using Terraria.GameContent.Drawing;
 using Terraria.ID;
 using Terraria.ObjectData;
+using Twaila.Config;
 using Twaila.ObjectData;
 
 namespace Twaila.Util
